+++ evaluation/experiments/libfuzzer-harness/.vscode/settings.json
@@ -1,0 +1,11 @@
+{
+    "files.associations": {
+        "*.json.liquid": "json",
+        "*.yaml.liquid": "yaml",
+        "*.md.liquid": "markdown",
+        "*.js.liquid": "liquid-javascript",
+        "*.css.liquid": "liquid-css",
+        "*.scss.liquid": "liquid-scss",
+        "istream": "c"
+    }
+}
